fixes review
