+++ lib/ui/menu_bottom_sheet.dart
@@ -1,8 +1,8 @@
-import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:pola_flutter/analytics/analytics_about_row.dart';
 import 'package:pola_flutter/analytics/pola_analytics.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'menu_bottom_item.dart';
 
 class MenuBottomSheet extends StatelessWidget {
   final PolaAnalytics analytics;
@@ -32,162 +32,93 @@ class MenuBottomSheet extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 11),
-            MenuBottomItem(
-              text: "O aplikacji Pola",
-              iconPath: 'info',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.aboutPola);
-                _launchURL('https://www.pola-app.pl/m/about');
-              },
-            ),
-            MenuBottomItem(
-              text: "O Klubie Jagielońskim",
-              iconPath: 'info',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.aboutKJ);
-                _launchURL(
-                    'https://klubjagiellonski.pl/o-klubie-jagiellonskim/');
-              },
-            ),
-            MenuBottomItem(
-              text: "Jak oceniamy Firmy",
-              iconPath: 'thumbs',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.instructionSet);
-                _launchURL('https://www.pola-app.pl/m/method');
-              },
-            ),
-            MenuBottomItem(
-              text: "Partnerzy",
-              iconPath: 'handshake',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.partners);
-                _launchURL('https://www.pola-app.pl/m/partners');
-              },
-            ),
-            MenuBottomItem(
-              text: "Przyjaciele Poli",
-              iconPath: 'diversity',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.polasFriends);
-                _launchURL('https://www.pola-app.pl/m/friends');
-              },
-            ),
-            MenuBottomItem(
-              text: "Oceń Polę",
-              iconPath: 'star',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.rateUs);
-                final url = Platform.isIOS
-                    ? "https://apps.apple.com/app/id1038401148"
-                    : "https://play.google.com/store/apps/details?id=pl.pola_app";
-                _launchURL(url);
-              },
-            ),
-            MenuBottomItem(
-              text: "Zespół",
-              iconPath: 'groups',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.team);
-                Navigator.pushNamed(context, '/web',
-                    arguments: "https://www.pola-app.pl/m/team");
-              },
-            ),
-            MenuBottomItem(
-              text: "Github",
-              iconPath: 'github',
-              onClick: () {
-                analytics.aboutOpened(AnalyticsAboutRow.github);
-                _launchURL('https://github.com/KlubJagiellonski');
-              },
-            ),
-            const SizedBox(height:23),
+            MenuBottomItems(analytics: analytics),
+            const SizedBox(height: 23),
             Container(
-              width: 296 + 32,
+              width: 328,
               height: 1,
               color: const Color(0xFFF0F0F0),
             ),
             const SizedBox(height: 17),
-            Row(
-              children: [
-                const SizedBox(width: 32.0),
-                Text(
-                  "Znajdź nas tutaj",
-                  style: const TextStyle(
-                    fontSize: 12,
-                    fontWeight: FontWeight.w700,
-                    fontFamily: 'Lato',
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 12),
-            SizedBox(
-              height: 60,
-              child: Row(
-                children: [
-                  const SizedBox(width: 32.0),
-                  Expanded(
-                    child: SingleChildScrollView(
-                      scrollDirection: Axis.horizontal,
-                      clipBehavior: Clip.none,
-                      child: Row(
-                        children: [
-                          HorizontalButton(
-                            text: "Twitter",
-                            onPressed: () {
-                              analytics.aboutOpened(AnalyticsAboutRow.twitter);
-                              _launchURL('https://twitter.com/pola_app');
-                            },
-                          ),
-                          const SizedBox(width: 14.0),
-                          HorizontalButton(
-                            text: "Facebook",
-                            onPressed: () {
-                              analytics.aboutOpened(AnalyticsAboutRow.facebook);
-                              _launchURL('https://facebook.com');
-                            },
-                          ),
-                          const SizedBox(width: 14.0),
-                          HorizontalButton(
-                            text: "Facebook",
-                            onPressed: () {
-                              analytics.aboutOpened(AnalyticsAboutRow.facebook);
-                              _launchURL('https://facebook.com');
-                            },
-                          ),
-                          const SizedBox(width: 14.0),
-                          HorizontalButton(
-                            text: "Facebook",
-                            onPressed: () {
-                              analytics.aboutOpened(AnalyticsAboutRow.facebook);
-                              _launchURL('https://facebook.com');
-                            },
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 17.0),
-            Row(
-              children: [
-                const SizedBox(width: 32.0),
-                Text(
-                  "Aplikacja Pola \n© Klub Jagielloński",
-                  style: const TextStyle(
-                      fontSize: 14,
-                      fontWeight: FontWeight.w600,
-                      fontFamily: 'Lato'),
-                ),
-              ],
-            ),
+            SocialMediaSection(analytics: analytics),
             const SizedBox(height: 33),
           ],
         ),
       ),
+    );
+  }
+}
+
+class SocialMediaSection extends StatelessWidget {
+  final PolaAnalytics analytics;
+
+  const SocialMediaSection({super.key, required this.analytics});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Row(
+          children: [
+            const SizedBox(width: 32.0),
+            const Text(
+              "Znajdź nas tutaj",
+              style: TextStyle(
+                fontSize: 12,
+                fontWeight: FontWeight.w700,
+                fontFamily: 'Lato',
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 12),
+        SizedBox(
+          height: 60,
+          child: Row(
+            children: [
+              const SizedBox(width: 32.0),
+              Expanded(
+                child: SingleChildScrollView(
+                  scrollDirection: Axis.horizontal,
+                  clipBehavior: Clip.none,
+                  child: Row(
+                    children: [
+                      HorizontalButton(
+                        text: "Twitter",
+                        onPressed: () {
+                          analytics.aboutOpened(AnalyticsAboutRow.twitter);
+                          _launchURL('https://twitter.com/pola_app');
+                        },
+                      ),
+                      const SizedBox(width: 14.0),
+                      HorizontalButton(
+                        text: "Facebook",
+                        onPressed: () {
+                          analytics.aboutOpened(AnalyticsAboutRow.facebook);
+                          _launchURL('https://facebook.com');
+                        },
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+        const SizedBox(height: 17.0),
+        Row(
+          children: [
+            const SizedBox(width: 32.0),
+            const Text(
+              "Aplikacja Pola \n© Klub Jagielloński",
+              style: TextStyle(
+                  fontSize: 14,
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'Lato'),
+            ),
+          ],
+        ),
+      ],
     );
   }
 
@@ -195,43 +126,6 @@ class MenuBottomSheet extends StatelessWidget {
     launchUrl(
       Uri.parse(url),
       mode: LaunchMode.externalApplication,
-    );
-  }
-}
-
-class MenuBottomItem extends StatelessWidget {
-  final String text;
-  final String iconPath;
-  final VoidCallback onClick;
-
-  const MenuBottomItem({
-    super.key,
-    required this.text,
-    required this.iconPath,
-    required this.onClick,
-  });
-
-  final textStyle = const TextStyle(
-      fontWeight: FontWeight.w500, fontSize: 16.0, fontFamily: 'Lato');
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onClick,
-      child: Column(
-        children: [
-          const SizedBox(height: 16),
-          Row(
-            children: [
-              const SizedBox(width: 32),
-              Image.asset("assets/icons/$iconPath.png"),
-              const SizedBox(width: 20.0),
-              Text(text, style: textStyle),
-              const SizedBox(width: 32),
-            ],
-          ),
-        ],
-      ),
     );
   }
 }
